Feat: finished adding an event
