+++ client/components/index.js
@@ -13,7 +13,7 @@ export {default as Channels} from './channels'
 export {default as SingleChannel} from './single-channel'
 export {default as VideoSession} from './videotest-kade'
 export {default as SearchResults} from './searchResults'
-export {default as EventsPage} from './events-page'
+export {default as EventsPage} from './EventPage'
 export {default as SearchUsers} from './SearchUsers'
 export {default as AddUser} from './AddUsers'
 export {default as UserProfiles} from './UserProfiles'
